Refactored widgets
`EventList` is now used in both places where events are shown (home and search screens)
